HRPV2-114 refactor getTopStudentsByActivityStatistics to use the same logic than getActivityStatistics to get number of activities or practice time

+++ oe-hr-portal-service-common/src/main/java/com/openenglish/hr/common/dto/PersonActivityTotalDto.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @Builder
 public class PersonActivityTotalDto {
     private PersonDto person;
-    private long totalActivities;
+    private double totalActivities;
 }
